Debounce text field input
