+++ Nexd/Nexd/utilities/swiftui/View+Nexd.swift
@@ -1,0 +1,29 @@
+//
+//  View+Accessibility.swift
+//  nexdScreenshots
+//
+//  Created by Tobias Schröpf on 27.05.20.
+//  Copyright © 2020 Tobias Schröpf. All rights reserved.
+//
+
+import SwiftUI
+
+extension View {
+    func identified(by identifier: AccessibilityIdentifier) -> some View {
+        accessibility(identifier: identifier.rawValue)
+    }
+
+    func cornerRadius(_ radius: CGFloat, corners: UIRectCorner) -> some View {
+        clipShape(RoundedCorner(radius: radius, corners: corners))
+    }
+}
+
+struct RoundedCorner: Shape {
+    var radius: CGFloat = .infinity
+    var corners: UIRectCorner = .allCorners
+
+    func path(in rect: CGRect) -> Path {
+        let path = UIBezierPath(roundedRect: rect, byRoundingCorners: corners, cornerRadii: CGSize(width: radius, height: radius))
+        return Path(path.cgPath)
+    }
+}
